requirements.txt
accelerate==1.10.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.2
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
dill==0.4.0
distro==1.9.0
docling==2.55.1
docling-core==2.48.4
docling-ibm-models==3.9.1
docling-parse==4.5.0
easyocr==1.7.2
en_core_web_sm==3.8.0
et_xmlfile==2.0.0
faiss-cpu==1.12.0
Faker==37.8.0
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.78.1
lazy_loader==0.4
llvmlite==0.45.1
lxml==5.4.0
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
murmurhash==1.0.13
narwhals==2.7.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numba==0.62.1
numpy==2.2.6
openai==2.1.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pillow==11.3.0
pip==25.2
plotly==6.3.1
pluggy==1.6.0
polyfactory==2.22.2
preshed==3.0.10
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydantic-settings==2.11.0
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.4
pynndescent==0.5.13
pyparsing==3.2.5
pypdfium2==4.30.0
pyphen==0.17.2
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
semchunk==2.2.2
sentence-transformers==5.1.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
textstat==0.7.10
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.9.30
tiktoken==0.11.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.0
typer==0.19.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
weasyprint==66.0
webencodings==0.5.1
wrapt==1.17.3
xlsxwriter==3.2.9
zopfli==0.2.3.post1
